requirements.txt
python-telegram-bot==20.7
ccxt
numpy
pandas
matplotlib
scipy
python-dotenv
requests
loguru
